mask_capsule in crop_aladin_pipe
